api/auth.py
"""
JWT authentication.

Phase 1: single-user credentials loaded from environment variables.
Phase 2 (multi-user): add a /api/auth/register endpoint that creates
User rows in the database and issues tokens per user.
"""

import os
from datetime import datetime, timedelta, timezone
from typing import Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from passlib.context import CryptContext
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY: str = os.getenv("SECRET_KEY", "insecure-default-change-me")
ALGORITHM: str = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES: int = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "1440"))

# ── Password hashing ──────────────────────────────────────────────────────────
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# ── OAuth2 bearer scheme — token expected in Authorization: Bearer <token> ────
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/auth/login")


def hash_password(plain: str) -> str:
    return pwd_context.hash(plain)


def verify_password(plain: str, hashed: str) -> bool:
    return pwd_context.verify(plain, hashed)


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    payload = data.copy()
    expire = datetime.now(timezone.utc) + (
        expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    payload["exp"] = expire
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)


def decode_token(token: str) -> dict:
    """Decode and validate a JWT. Raises HTTPException on failure."""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid or expired token",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
        return payload
    except JWTError:
        raise credentials_exception


async def get_current_user(token: str = Depends(oauth2_scheme)) -> str:
    """FastAPI dependency — returns the username from a valid JWT."""
    payload = decode_token(token)
    return payload.get("sub")


# ── Phase 1: single-user credential check ─────────────────────────────────────

def authenticate_single_user(username: str, password: str) -> bool:
    """
    Validates against DASHBOARD_USERNAME / DASHBOARD_PASSWORD env vars.
    In Phase 2, replace this with a database lookup.
    """
    expected_username = os.getenv("DASHBOARD_USERNAME", "admin")
    expected_password = os.getenv("DASHBOARD_PASSWORD", "changeme")

    if username != expected_username:
        return False

    # Support both plain-text passwords (dev) and bcrypt hashes (prod).
    # If the env var starts with '$2b$' it's already bcrypt-hashed.
    if expected_password.startswith("$2b$"):
        return verify_password(password, expected_password)
    return password == expected_password
